A lot
Made questions clickable in all tables. Changed Database setup so that
you can’t duplicate users and you can link questions and answers
together. And users to those questions and answers. Fixed the UI for
the answer question page. Updated the Edit Profile screen so that the
tables are populating with the correct information.